Switch on parallel processing as per Readme.txt
"To run the code on a parall computer with MPI, the user has to comment out the line "use mpistub" in Contrl/Input.f90, DataStruct/Data.f90, DataStruct/Pgrid.f90, DataStruct/PhysConst.f90, and Func/Timer.f90.
The user also has to remove the mpif.h file under the Appl, Control, DataStruct, and Func directories.
The user also has to modify the Makefile to remove the mpistub.o inside the file and to use the appropriate parallel F90 compiler such as mpif90."