Multi-bunch plotting capabilities
Adds three plots with multiple bunches:
* Beam size (rms)
* Emittance
* Emittance growth

Plots each bunch as a dotted line, and a combined value for all bunches as a solid line.

Can include all bunches or only a selected number of bunches.

Also plots experimental data for beam size alongside the simulation results.

MultiBunchPlot.py can also be run as a script, outputting all files as PNG images, and can be run on a headless server.
The number of bunches to plot can be given as a command-line parameter.