New method of getting number of bunches
Previously counted the number of input files, but now there may be an `ImpactT10.in` input file even if the electron bunch is not used.

Now reads the number from the second input line of the first input file, which is the same as how _Impact-T_ itself gets the number.