Add option to include a z-offset in multi-bunch plots
Sometimes an offset to the starting position is required to ensure no particles have negative starting position. In such cases, the z-location of particles should be adjusted to enable direct comparison of experiment and simulation.

Specifying a z-offset makes the plots as if the particles had started at z=0 rather than at the offset location.